refs #409 - add site ID to cookie name; shorten domain hash to 16 bits (4 hexit characters)
This is a hybrid between the previous implementation and what I proposed.
 * Adding idsite to the cookie name means subdomains that track using different site IDs can still use/share subdomain cookies
 * Keeping the domain hash in the cookie name will make it easier in future to delete invalid cookies (integrity check)

Decided not to auto-set www.example.com's cookie domain=.example.com -- as the convenience introduces side-effects, and I have a feeling will be more trouble than beneficial.  Will continue to leave it to the user to explicitly set the cookie domain.  Users should be advised to redirect example.com to www.example.com (or vice-versa) to:

a) to avoid separate cookies between the two domains, and
b) to improve SEO.  (Google for "seo www vs no www".)