refs #2465
overlay has two modes now: full screen with sidebar & full screen without sidebar. for now, it is only accessible via a new row action in the pages report.

because of this change, some mostly not-so-nice core-changes could be removed (mainly that's good because overlay doesn't impact as many things outside the plugin anymore):

 * disableHistory option of broadcast
 * handling for parameter overlayUrl in broadcast
 * openLinksInOverlay() method of html table
 * config option overlay_launch_from_pages_report
 * hack for data table to launch overlay from the links

other changes:

 * error message for wrong domain has nice styles and links to admin page when the user has admin privileges
 * when registering row actions, the order can be set
 * made row action style a little more compact
 * increased min width of label column a bit

git-svn-id: http://dev.piwik.org/svn/trunk@7470 59fd770c-687e-43c8-a1e3-f5a4ff64c105