refs #4009 making the number of rows after which the PDFReports plugin truncates configurable.
this does not fix the ticket since it still would be nice if the truncation would be a little more intelligent. i.e. using different values for landscape / portrait and/or not truncating at all in HTML view.

julien, if you disagree, feel free to close the ticket. it's just a suggestion and i fixed the part that caused trouble for me so i'm fine either way.