fixing PDFReport bug with period=rage
when a range is selected in the date picker and the download link in the pdf report admin section is used, the currently selected date and the period of the report are put together in one url. this results for example in period=week and date=2013-05-01,2013-05-23. this causes errors because the API then returns a DataTable_Array instead of regular DataTable. to fix this, the download url should only contain a single date (the start date of the selected range).