Applying change suggested by travis
npm is included with Node, so there is no separate package for it. Our Node (and npm) is installed with nvm, and as such is not directly available with sudo. Changing the last few lines of the install-ubuntu.sh script to this should work (and removing the install of the npm package):

npm config set loglevel error
npm install .