Remove CacheFile cache from APC on write
If APC is enabled on the piwik server, cache files written by the CacheFile
class are stored in the opcode cache. This causes disk io problems when doing
things like loading weblogs into piwik with the import_logs.py script,
especially when running in --disable-bulk-tracking mode. In this scenario,
apache write IO quickly reaches 100% and slows the import to a crawl.