refs #6415 better workaround for possible empty visits
We will now always wait 800ms after a first tracking request was issued.
This should give enough time to create a visitor before any other request
is executed. Ideally we will resolve this issue on the server side as
this problem can occur using other SDKs as well and for some servers or
sometimes the 800ms might be not long enough.