refs #6435 fix some more wrong values if a German locale is used
This caused me a lot of headache. I ran the system tests with a German
locale and noticed so many values are completely different. I was looking
for places where there were still any float to string conversions and I
was wondering why the existing fix didn't work until I noticed the
PiwikTracker did not send floats but floats with a comma as decimal
point which in the end was casted to ints etc.