Removed group check before running any PHP tests.
This check does not really makes sense as a group was never needed
for a test to be executed on Travis. The testsuites that always have been
used on Travis automatically execute all tests within the PHPUnit/Unit
Integration and System folder. A group was never needed. Maybe some
developers used groups to execute tests but then always only a part of
the tests was executed as a test does not necessarily has the "correct"
group. Removing this check is also important for plugin developers who
write tests for their plugin as we cannot expect that they set a group
as it is usually not needed in any other project. If plugin developers
do not set a group and we run our automated tests no test
would be executed because of this unneeded check.