Update piwik.js
Tried implementing sendBeacon on the client side, but it seems the server side requires a different content-type? Remove "return false" from 2421 for testing.

In my experience the 500ms wait is arbitrary guesswork and is not enough for slow connections from distant places. I increased this in my own implementation to 5000ms and only use it as a fallback timeout so that the navigation do not get stuck forever.

Also the image.onready now requires image.onerror as well since piwik does not return a valid image anymore (return code 204)