rename Admin menu item Settings => Plugins to Settings => Plugin Settings.
It is too confusing to have "Plugins" menu item twice in the menu and
the same menu item is called "Plugin Settings" in User menu already.
We should have consistent naming in User menu and Admin menu for this.
It is a bit annoying to have "Settings" duplicated as it is already
mentioned in the category name but there are more advantages than
disadvantages