refs #6811 make sure the endDate is valid when switching between websites
Imagine website is created on 2015-12-20, if current selected date
was eg 2015-04-18,2015-04-29 the startDate was automatically fixed
already to 2015-12-20,2015-04-29 but not the endDate. This is now
done and it will apply 2015-12-20,2015-12-20.
The code looks a bit more complex as there was another bug that
although it applied 2015-12-20,2015-12-20 it did still
display the date range for 2015-04-18,2015-04-29 in the UI in
the calendar which is wrong. Also I do now fix the date in case
period != range. So period=day&date=2015-04-29 would result in
period=day&date=2015-12-20