refs #6705 Hide users metric in case the feature is not used.
This is an alternative solution to fix this problem. The difference
here is that we only do an additional call to VisitsSummary::get
if users is 0. The disadvantage is that eg API.getReportMetadata
will report that nb_users exist but the API.getProcessedReport
might not return it. This already caused some problems while
trying to implement it like this as API.getProcessedReport will
add all metrics returned by getReportMetadata automatically. So even
if we remove nb_users in the API output of VisitsSummary.get it
would be still added again. Another disadvantage of this solution
is that the code is more fragile as we have to map parameters
like $params[0] to idSite, $params[1] to period, ...