refs #7104 This should prevent from accidentally starting a Zend_Session.
Sometimes, eg when module=API, we do not start a session. If any code
part during such a request uses SessionNamespace although we did not start
a session, Zend_Session_Namespace will start a Zend_Session which does not
respect our configuration. It would use default Zend Session options
and therefore probably file based sessions. Therefore, we force the
creation of a "Piwik session". If we have created the session before it
won't be started again.