#6441 Updating through HTTPS if openssl is installed, else use HTTP
HTTP will be used if (and only if) HTTPS is not supported at all (e.g. openssl is not installed or fopen/curl is not supported). If HTTPS is supported, we try to update through HTTPS: any failure will stop the update process and we don't fallback to HTTP. This is the expected behavior: a failure could be a MITM attack.