fix various issues with user menu.
* When clicking on the anonymous user a login form was shown
* Some menu items were shown although the anonymous user has no access to it

We will now load the Help page when clicking on anonymous user name.
If the plugin Feedback is not activated we will open the API page since
this has to be activated. If Feedback plugin is disabled the anonymus user login
which will be a known issue but I think we can ignore this case.