Change private variables and methods to protected.
The current use of "private" prevents class variable and method inheritance in child classes. The "protected" ones on the other hand, offer the same access control, but can be inherited.

Impact: I'm working on a plugin to encrypt the transmitted password. That plugin "extends" many of the classes of the core Login plugin. The current usage of "private" forces me to, unnecessarily duplicate (copy-paste) the "private" members of the Login parent class, without any modification. I tested this patch with 2.14.3 and it works flawlessly with my (future) plugin.

Also, it maybe could help https://github.com/torosian/LoginRevokable/issues/1.

Thanks in advance.