Piwik\Plugins\Login\Controller class: Change private members to protected
To the Piwik\Plugins\Login\Controller class: This change is to prevent extending classes (third-party login plugins reusing the core Login plugin implementation) from having to copy-paste \Login\Controller code which, if declared "private" cannot be part of the child class.

Please refer to https://github.com/piwik/piwik/pull/8681 for details - that PR, which is an exact copy, was merged into the Piwik 2.x branch, but then reverted to not break existing plugins. It was decided to do that on the 3.x branch instead, which this PR does.