Fix checking of $HTTP_RAW_POST_DATA for PHP7
Due to populating of raw post data being deprecated in PHP5.6 We
needed to set the ini value `always_populate_raw_post_data` to
`-1` in order to avoid deprecation warnings and the like. However
this feature has been removed in PHP7. The knock on to this is
that regardless of wether the `php.ini` contains the line
`always_populate_raw_post_data=-1` the `ini_get` function evaluates
to 0/false. Thus we only need to check for this ini setting being
-1 for PHP5.6