Improve extendability of userSettings template
Done in order to make it easier for child templates to manipulate just the settings table part of the template. For the core UsersManager plugin itself, on the other hand, addition of the new "block" tag has no impact at all. It's just a marker which can be used by child templates if desired.