Remove depedency to SitesManager plugin in core ArchiveInvalidator service by moving logic to new request processor.
ArchiveInvalidator used to update site creation time. The reasoning was, if we tracked data in the past that was before
the site's creation time, the site creation time should be updated, so it will be displayed in the UI. The tracking
should trigger archive invalidation which should eventually trigger an update to the site creation time. This is
incorrectly mixing concerns. Now the new request processor will make this change at the moment old data is tracked.

New test included.