Reset the list of items in the cart, after tracking an ecommerce conversion (#10279)
* Clarify that token_auth must be  an admin or super user token.

* Reset the list of items in the cart, after tracking an ecommerce conversion

But do not reset the list of itms in the cart after tracking an ecommerce cart update.
fixes #10252

* Adding changelog note
`piwik.js`: after an ecommerce order is tracked using `trackEcommerceOrder`, the items in the cart will now be removed from the JavaScript object. Calling `trackEcommerceCartUpdate` will not remove the items in the cart.