Enforce the Minimum PHP version in composer.json using "platform" feature Added in https://github.com/composer/composer/commit/a57c51e8d78156612e49dec1c54d3184f260f144
It was found that this flag is very important to use to make sure the package we build and release will be compatible with this PHP version.
Otherwise, the minimum PHP version enforced would be the version where our build script runs, which could be PHP 5.5 or PHP 7 which would have caused serious issues.