Trigger JS tracker update when checking for updates
If for some reason the tracker file does not get updated, we will regenerate it immediately when someone is checking for updates manually. This way we can tell users to perform this action to force regeneration immediately (if there was a change). This way users don't have to log in to their server and force regeneration via a command when it was not updated automatically.