make sure to ignore free plugins (#10940)
This is related to https://forum.piwik.org/t/license-missing-of-a-free-glp-plugin/21674 

Some users see a license warning for a free GPL plugin. I have spent hours trying to understand how this could possibly happen over the last weeks but not sure how it happens. It must be some kind of caching issue with a certain PHP version or something. This change might not fix the actual issue but is a good check to have on top to check for invalid licenses only if it is not a free plugin.