Update code formatting for better readability
This is a little bit crazy but apparently the only way I can answer
your questions on my PR are via this commit message.

Apparently the owner of the piwik repo restricted comment to only
prior contributor. You would think that this would allow you to
comment your own PR but apparently not.

So abouts yours questions:

In my case the archiving tasks is run by a cron job via
```./console core:archive```, because of that it's quite
convenient to see in the output if it's the memory
constraints that made the cronjob fail.

For the conformance to the code format I hope that this commit fix it.
Don't hesitate to point me other formatting issue if there are others.

Thanks