Improve serialize dataTable performance (#12289)
* Improve serialize dataTable performance

This seems to improve the performance of serializing a data table which is useful when archiving. In my case it made the archiving of a monthly archive twice as fast from about 120 seconds down to like 70seconds.

* remove no longer needed variable

* faster check whether an array key is set

It is known that array_key_exists is quite a bit slower than `isset`. As this is executed easily a few million times this makes a difference in performance but here `array_key_exists` and `isset` work the same in logic as we never set `$consecutiveSubtableIds[$id]=null`

* Assignment should be no longer needed as passing by reference