Improve detection of piwik/matomo URL
refs a couple of issues I think. Eg https://github.com/matomo-org/matomo/issues/11029 

For some reason that I cannot explain, it stored `127.0.0.1` as the matomo URL in the DB causing a bug in tag manager. So I opened a regular page in Matomo and was expecting it to correct the value in the option table but it didn't. I think this is because of the above change where it checks whether the previous URL is not a local host whereas it maybe should check on the current URL? 

See the debugger picture where it did not overwrite 127.0.0.1 with the correct current url/host because the host is currently a local host (127.0.0.1). If it checked whether the actual host is a local host, it would have updated it.