Visit by hour should show report based on first action time
Just a suggestion. IMO it is way more important to see at what hour you acquired specific users compared to when they dropped off. Especially if you are running certain ads for example etc. We could also add another report "first" and "last action" hour but I think that not both are actually needed and keeping it simple be better.

Performance wise this shouldn't have really any impact as the index be still used in the back by the where statement.