Do not apply local config in tests
I often have tests failing because I have in my local `config/config.php` a DI value overwritten which is set in `plugins/*/config/test.php` or `plugins/*/config/config.php` for example. I was thinking a quick fix by to ignore the local config file when executing tests which is I presume similar to what we do with the config ini file?