Tooltips appear all over the place but shouldn't (#13442)
* Tooltips appear all over the place but shouldn't

fix #13305 

Add a delay so tooltips don't appear immediately. Do not show any tooltip for any form field.

In visitor log the titles are still shown immediately but there it is wanted maybe?

I clicked through couple of pages and looked a bit in the code but it is hard to find all the places where a title attribute should be removed. Maybe we do this better over time.

* Remove tooltip as the link already shows the same text