Start delete log task hourly (#13824)
This allows plugins to better randomise when a log deletion task should be executed by eg overwriting config ( eg setting `delete_logs_schedule_lowest_interval = "7.1"`.

It won't execute the task any more often otherwise by changing it to hourly because the `deleteLogData` checks when it was last executed. It's super fast this check (reading an option entry) so there is no performance problem by that.