handle background color of opt out if font is almost white
If no background color is set, the background color is transparent. But the whole site background color is white, so if you choose white or a color that is nearly white as font color the iframe appears so be empty.

This commit implements a new function that tests if a hex color is nearly white and if yes the code will change the background color to a grey color.