When dispatch is disabled via a constant, it should not be dispatched
Eg when Matomo is not installed, it would still dispatch the request in https://github.com/matomo-org/matomo/blob/3.8.1-b1/plugins/Installation/Installation.php#L111 even when `PIWIK_ENABLE_DISPATCH` is disabled. 

Will set it to WIP for now as I'm not sure if we want to have this actually merged or not.