Let plugins change the report type in sendReport event (#14098)
This way I can set the `reportType` to eg `foobar` so it won't be handled and won't be sent. Needing this feature... Was going to add otherwise another parameter or another event but having two events with similar `sendReport` name may be confusing and another parameter could cause issues. No need to document this any further I would say.

@diosmosis can you have a quick look?