Show correct message why plugin was deactivated when it has a missing… (#14378)
* Show correct message why plugin was deactivated when it has a missing license

I was never able to reproduce this issue until I disabled development mode. 
With the old code, when no valid license in the Marketplace was configured and development mode is disabled, it would show a message "Plugin XYZ could not be loaded" where it should have said "Plugin was not loaded because the license is missing".

This was regressed in https://github.com/matomo-org/matomo/blob/3.10.0-b1/core/Plugin/Manager.php#L1058-L1086 where we unloaded plugin when the license is missing, so that logic failed to determine that the plugin was activated. By looking at the config we can ensure to correctly detect if plugin is still activated or not and now shows correct message.

A test is not really easily possible to write as some logic is disabled for tests etc and might not be actually needed.

* rename method to be more clear