Check $defaultLayout is an array when generating default dashboard (#14596)
fix https://github.com/matomo-org/matomo/issues/14595

Not sure why it wouldn't be an array. Even when using ` json_decode($defaultLayout);` compared to ` json_decode($defaultLayout, true);` it is an array on my instance and works nicely. Just in case there is some other plugin installed that changes it maybe, check if it is an array