Don't show feedback prompt when user does not exist (#14692)
Prevents an error in the UI that said

```
Date format must be: YYYY-MM-DD, or 'today' or 'yesterday' or any keyword supported by the strtotime function (see http://php.net/strtotime for more information):
``` 

when the user did not exist. Eg when you have "virtual" users that aren't actually existing in the user table.