Fix error session already started
Got this error when going on eg  `index.php?module=PrivacyManager&action=privacySettings&idSite=1&period=week&date=2019-09-20` and a session was started by another tool. That page uses NONCE which then uses SessionNamespace which then wants to start the session even though it was already started.
I don't think I can write a test for it but need to see.